Remove NewRelic, add Docker setup
NewRelic never alerted on any of Spud's failures, so either I used it
wrong, or it just wasn't right for my application. In either case, it's
gone now.

Added a first-cut Dockerfile to see if we can start deploying everything
via Docker instead of dealing with stepping on each other's toes.